// src/app/Model/Configuration/Userclient.ts
export class Userclient {
    id: String;
    clientid: Number;
    clientname: String;
    subclientcode: String;
    subclientname: String;
    userid: String;
    username: String;
    firstname: String;
    lastname: String;
    createdby: String;
    createdon: String;
    groupid: String;
    status: boolean;
}